include option to have no fourier series; check to make sure poly is at least 1
if some part of fourier series is required, then you cannot use this function if the time series has a frequency of 1 or less. Originally I thought this would not be a big deal b/c then simpler functions (e.g., based on just trend_xreg along) could be used. But considering that I want to use 1 generic function to apply to a large number of time series, this becomes problematic. Specifically, I want to detrend time series sampled at different time scales (frequencies); these include 5-min resolution up to 2-day resolution, and the period of the parent time series (5-min data) is 1 day, such that some of these time series have a frequency of 1 or less. Thus, the function had to be more robust.