increment to v0.2 -- eco simulation, ar(p), tvarss
Lots of work done to include an ecological model of lake eutrophication that could be simulated to validate statistical methods. Furthermore, the addition of functions supporting ar(p) and tvarss is significant.

Next version will include applying the old analyses (rolling window ar(1)) to simulated data, as well as applying new statistics (ar(p) and tvarss) to both field data and to the ecological simulation. In this process, I expect the recently added functions to be adjusted, and complementary functions will likely be added. Nonetheless, the current skeleton for these steps is a good milestone.