move simLakeP_Data to report script
original purpose of simLakeP_Data.R was to simulate a lake P time series and save the output as a .RData to save as part of the package. I then realized that for a simulation of modest size, there was no point to this if I wrote functions that made the code for doing the simulation sufficiently concise --- this I accomplished. The script later grew to include analysis of the simulated data, which is beyond the scope of the script (and therefore the script's name no longer reflected its full function). As the analyses grew, I knew I was movign beyond this scope and intended to move the contents elsewhere. I moved the conents into timeScales_dlm_arp_sim.R to accurately reflect its broad purpose, and also formatted that script as a 'notebook' to produce a .html report

major revisions from the original script to the report include: 1) altering axis labels and margins a bit (don't label x-axis for all rows); 2) do NOT take the log of the time series before analysis; 3) the tvReg AR(p) model w/o detrending now has a time-varying mean